Allow to specify if widgets are resized when calling setGridWidth
I added a argument 'doNotPropagate' to the setGridWidth. If this argument is explicitly passed as true the inner widgets will not be resized.

And In the setGridWidth method i also updated the GridstackEngine 'width' attribute in addition to the opts.width. Il you do not do it you will not be able to move/resize widgets to the the new space generated if the parent grid has been expanded.