Bug wrong maxHeight for updateStyles function
When there are 4 row of grid items and more. There are bug of updates styles. 
The function does not take into account all nodes, but only those that have been changed (_dirty).
In this regard, the styles for the lower layers do not sink, and they rise upward.
The solution is to calculate the maxHeight for all nodes.