(fix): moved the `dropped` event handler call to a place where the drop is completely done.
In rare cases the call to the `dropped` handler could lead to errors.

The error occurred in the `window.setTimeout` code right after the call. An example is when one executes some logic on the (droppedCB) event in Angular. If you reset the source grid the item was dragged from (by resetting the items in an *ngFor), the grid crashed.

By executing the event in the `setTimeout` this problem is solved. It also makes logically sense as the event noting the drop is done is now fired after `this._prepareDragDropByNode(node);` is executed.