Move the item down to the first available position
First look for the next item in the row directly below the selected
item. When nothing is find look for the first item below the selected
item. And when that fails because of whitespacing, look for the first
item which is below and overlaps the selected item.

Also check if the selected item has sibling, because when the item below
spans more columns, it can mean there is no space for that item.
Instead, move the selected item below the next item.